+++ nab3/aws.py
@@ -7,7 +7,6 @@ class AWS(BaseAWS):
 
     def __init__(self, session: boto3.Session = boto3.Session()):
         self._client = ClientHandler(session)
-        self._region = session.region_name
 
     def __getattr__(self, value):
         if value in self._service_map.keys():
@@ -20,7 +19,3 @@ class AWS(BaseAWS):
         :return:
         """
         return sorted([k for k in self._service_map.keys()])
-
-    @property
-    def region(self):
-        return self._region