+++ nab3/service/ecs.py
@@ -1,6 +1,6 @@
 import logging
 
-from nab3.mixin import AppAutoScaleMixin, AutoScaleMixin, MetricMixin
+from nab3.mixin import AppAutoScaleMixin, AutoScaleMixin, MetricMixin, SecurityGroupMixin
 from nab3.base import BaseService
 
 LOGGER = logging.getLogger('nab3')
@@ -111,7 +111,7 @@ class ECSInstance(BaseService):
     )
 
 
-class ECSCluster(AutoScaleMixin, MetricMixin, BaseService):
+class ECSCluster(AutoScaleMixin, MetricMixin, SecurityGroupMixin, BaseService):
     """
     boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ecs.html#ECS.Client.list_clusters
     boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ecs.html#ECS.Client.describe_clusters
@@ -130,6 +130,18 @@ class ECSCluster(AutoScaleMixin, MetricMixin, BaseService):
         self.create_service_field('instances', 'ecs_instance')
         self.create_service_field('services', 'ecs_service')
         super(self._get_service_class('ecs_cluster'), self).__init__(**kwargs)
+
+    async def get_on_demand_monthly(self, currency='usd'):
+        if not self.asg.is_loaded():
+            await self.fetch('asg')
+
+        return await self.asg.get_on_demand_monthly(currency)
+
+    async def get_on_demand_hourly(self, currency='usd'):
+        if not self.asg.is_loaded():
+            await self.fetch('asg')
+
+        return await self.asg.get_on_demand_hourly(currency)
 
     async def load_asg(self):
         """Retrieves the instances asg.
@@ -178,6 +190,14 @@ class ECSCluster(AutoScaleMixin, MetricMixin, BaseService):
             return self.instances
         self.instances = await self.instances.list(cluster=self.name)
         return self.instances
+
+    async def load_security_groups(self):
+        if self.security_groups.loaded:
+            return self.security_groups
+
+        await self.fetch('asg__security_groups')
+        self.security_groups = self.asg.security_groups
+        return self.security_groups
 
     async def load_services(self):
         """Retrieves the cluster's services.