+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ---
 
+# 0.1.1 (2020/09/28)
+
+### Bug Fixes
+* Cleaned up README
+* Added an interval argument to `set_service_stats` and `set_n_service_stats`
+
+### Features
+* Added support for `ECSCluster.security_groups`
+* Created a `PricingMixin`
+* Pricing support added to:
+  * `EC2Instance` 
+  * `ASG` 
+  * `ECSCluster` 
+
+---
+
 # 0.1.0 (2020/09/22)
 
 ### Bug Fixes